[LB-462] feat : 트레이싱 설정
